#255 [FEAT] 알림 즉시 발송/ 예약 발송 api Request DTO 구현
-바뀐 기능에 맞게 사용하지 않는 클래스는 제거하고, 즉시 발송과 예약 발송 Request DTO를 구현했습니다.
