--- operation-api/src/main/java/org/sopt/makers/operation/web/alarm/dto/request/AlarmSendRequest.java
@@ -1,6 +1,0 @@
-package org.sopt.makers.operation.web.alarm.dto.request;
-
-public record AlarmSendRequest(
-		long alarmId
-) {
-}
